Split web API client folders
