+++ apps/web/src/api/endpoints/progress.ts
@@ -1,9 +1,9 @@
-import { ApiContractError } from "../apiContracts/core";
+import { ApiContractError } from "../../apiContracts/core";
 import {
   parseProgressReviewScheduleResponse,
   parseProgressSeriesResponse,
   parseProgressSummaryResponse,
-} from "../apiContracts/progress";
+} from "../../apiContracts/progress";
 import type {
   ProgressReviewSchedule,
   ProgressReviewScheduleInput,
@@ -11,9 +11,9 @@ import type {
   ProgressSeriesInput,
   ProgressSummaryInput,
   ProgressSummaryPayload,
-} from "../types";
-import { parseContractResponse } from "./response";
-import { allowAuthRecoveryWithTransientNetworkRetry, requestJson } from "./transport";
+} from "../../types";
+import { parseContractResponse } from "../transport/response";
+import { allowAuthRecoveryWithTransientNetworkRetry, requestJson } from "../transport/transport";
 
 export async function loadProgressSummary(input: ProgressSummaryInput): Promise<ProgressSummaryPayload> {
   const searchParams = new URLSearchParams({
